requirements.txt
openai==0.28.0
python-dotenv==1.0.0
jieba==0.42.1
sentence-transformers==2.2.2
gradio==3.50.0
scikit-learn==1.2.0
pandas==1.5.0
matplotlib==3.7.1
numpy==1.24.3 